requirements.txt
python==3.11.6
torch==2.0.1
transformers==4.35.0
sentencepiece==0.1.99
numpy==1.26.0
fastapi==0.103.0
uvicorn==0.23.2
paho-mqtt==1.6.1
pyautogui==0.9.54
librosa==0.10.1
speechrecognition==3.10.0
gTTS==2.3.2
opencv-python==4.8.1
diffusers==0.21.4
bitsandbytes==0.41.1
pandas==2.1.1
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2
networkx==3.1
sentence-transformers==2.2.2
qiskit==0.44.1
web3==6.11.0
matplotlib==3.8.0
seaborn==0.13.0
music21==9.1.0
cryptography==41.0.4
langchain==0.0.300
chromadb==0.4.10
boto3==1.28.0
moviepy==1.0.3
peft==0.5.0
scikit-learn==1.3.0
axios==1.6.0